refactor: RoomescapeSchedule 패키지 이동
