+++ src/main/java/roomescape/domain/reservation/Reservation.java
@@ -14,6 +14,7 @@ import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
+import roomescape.domain.RoomescapeSchedule;
 import roomescape.domain.user.User;
 import roomescape.exception.BusinessRuleViolationException;
