+++ src/main/java/roomescape/domain/reservation/ReservationQueues.java
@@ -9,6 +9,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import roomescape.domain.RoomescapeSchedule;
 
 public class ReservationQueues {
