+++ src/test/java/roomescape/TestFixtures.java
@@ -3,9 +3,9 @@ package roomescape;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.concurrent.atomic.AtomicLong;
+import roomescape.domain.RoomescapeSchedule;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationStatus;
-import roomescape.domain.reservation.RoomescapeSchedule;
 import roomescape.domain.theme.Description;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeName;
