+++ src/test/java/roomescape/application/TimeSlotServiceTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
+import roomescape.domain.RoomescapeSchedule;
 import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.RoomescapeSchedule;
 import roomescape.exception.InUseException;
 
 @Import(TimeSlotService.class)
