+++ src/test/java/roomescape/domain/reservation/ReservationRepositoryTest.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import roomescape.TestRepositoryHelper;
+import roomescape.domain.RoomescapeSchedule;
 import roomescape.exception.NotFoundException;
 
 @DataJpaTest
